Remove manager app, create manager-expo app
Using Expo (create-react-native-app) instead of react-native-cli for the manager app.
Reason being, this app uses an old version of react-native-router-flux, which doesn't seem to work with the latest version of react-native, meaning I have to use an old version of react-native.
If I use an old version of react-native, I cannot specify the port in `react-native run-ios`, and it will default to 8081.
This port is used on my corporate laptop, meaning the app will not run, as the port is in use.

Expo solves this by using a different port by default.